requirements.txt
Flask~=3.1.0
gunicorn
grok3api~=0.0.1a14
httpx~=0.28.1
python-dotenv~=1.0.1